+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/DependencyGraphTransformer.kt
@@ -1309,31 +1309,3 @@ internal class DependencyGraphTransformer(
     companionObject.dumpToMetroLog()
   }
 }
-
-internal class GraphGenerationContext(
-  val graph: IrBindingGraph,
-  val thisReceiver: IrValueParameter,
-  // TODO we can end up in awkward situations where we
-  //  have the same type keys in both instance and provider fields
-  //  this is tricky because depending on the context, it's not valid
-  //  to use an instance (for example - you need a provider). How can we
-  //  clean this up?
-  val instanceFields: Map<IrTypeKey, IrField>,
-  val providerFields: Map<IrTypeKey, IrField>,
-  val multibindingProviderFields: Map<Binding.Provided, IrField>,
-  val bindingStack: IrBindingStack,
-) {
-  // Each declaration in FIR is actually generated with a different "this" receiver, so we
-  // need to be able to specify this per-context.
-  // TODO not sure if this is really the best way to do this? Only necessary when implementing
-  //  accessors/injectors
-  fun withReceiver(receiver: IrValueParameter): GraphGenerationContext =
-    GraphGenerationContext(
-      graph,
-      receiver,
-      instanceFields,
-      providerFields,
-      multibindingProviderFields,
-      bindingStack,
-    )
-}
